split position embed code from MHA
